Move local/offproc service interfaces out of platform
Part of #127878
