+++ src/vs/workbench/contrib/terminal/electron-sandbox/terminal.contribution.ts
@@ -6,11 +6,11 @@
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { registerSharedProcessRemoteService } from 'vs/platform/ipc/electron-sandbox/services';
 import { Registry } from 'vs/platform/registry/common/platform';
-import { ILocalTerminalService, TerminalIpcChannels } from 'vs/platform/terminal/common/terminal';
+import { TerminalIpcChannels } from 'vs/platform/terminal/common/terminal';
 import { ILocalPtyService } from 'vs/platform/terminal/electron-sandbox/terminal';
 import { IWorkbenchContributionsRegistry, Extensions as WorkbenchExtensions } from 'vs/workbench/common/contributions';
 import { ExternalTerminalContribution } from 'vs/workbench/contrib/externalTerminal/electron-sandbox/externalTerminal.contribution';
-import { ITerminalProfileResolverService } from 'vs/workbench/contrib/terminal/common/terminal';
+import { ILocalTerminalService, ITerminalProfileResolverService } from 'vs/workbench/contrib/terminal/common/terminal';
 import { LocalTerminalService } from 'vs/workbench/contrib/terminal/electron-sandbox/localTerminalService';
 import { TerminalNativeContribution } from 'vs/workbench/contrib/terminal/electron-sandbox/terminalNativeContribution';
 import { ElectronTerminalProfileResolverService } from 'vs/workbench/contrib/terminal/electron-sandbox/terminalProfileResolverService';
